+++ controls/urls.py
@@ -1,8 +1,14 @@
 from django.conf.urls import include, url
 
+# from django.conf.urls import patterns, include
+# urlpatterns = urlpatterns + patterns('', (r'^', include('urls_api')))
+
 from django.contrib import admin
 from django.conf import settings
 admin.autodiscover()
+
+from django.views.decorators.csrf import csrf_exempt
+from . import urls_api
 
 import controls.views
 from . import views
@@ -33,6 +39,13 @@ urlpatterns = [
     url(r'^statement_history/(?P<smt_id>.*)/$', views.statement_history, name="statement_history"),
     url(r'^restore_to/(?P<smt_id>.*)/(?P<history_id>.*)/$', views.restore_to_history, name="restore_to"),
 
+    # Systems Assessment Results
+    url(r'^(?P<system_id>.*)/assessments$', views.system_assessment_results_list, name="system_assessment_results_list"),
+    url(r'^(?P<system_id>.*)/assessment/new$', views.manage_system_assessment_result, name="new_system_assessment_result"),
+    url(r'^(?P<system_id>.*)/sar/(?P<sar_id>.*)/view$', views.view_system_assessment_result_summary, name="view_system_assessment_result_summary"),
+    url(r'^(?P<system_id>.*)/sar/(?P<sar_id>.*)/edit$', views.manage_system_assessment_result, name="manage_system_assessment_result"),
+    url(r'^(?P<system_id>.*)/sar/(?P<sar_id>.*)/history$', views.system_assessment_result_history, name="system_assessment_result_history"),
+
     # Systems Inventory and Deployments
     url(r'^(?P<system_id>.*)/deployments$', views.system_deployments, name="system_deployments"),
     url(r'^(?P<system_id>.*)/deployment/new$', views.manage_system_deployment, name="new_system_deployment"),
@@ -49,8 +62,8 @@ urlpatterns = [
     # System Components/Elements
     url(r'^(?P<system_id>.*)/components/selected$', views.SelectedComponentsList.as_view(), name="components_selected"),
     url(r'^(?P<system_id>.*)/components/selected/export/opencontrol$', views.export_system_opencontrol, name="export_system_opencontrol"),
-    url(r'^(?P<system_id>.*)/component/(?P<element_id>.*)/download/oscal/json$', 
-        views.system_element_download_oscal_json, 
+    url(r'^(?P<system_id>.*)/component/(?P<element_id>.*)/download/oscal/json$',
+        views.system_element_download_oscal_json,
         name="system_element_download_oscal_json"),
     url(r'^(?P<system_id>.*)/component/(?P<element_id>.*)$', views.system_element, name="system_element"),
     url(r'^(?P<system_id>.*)/controls/updated$', views.controls_updated, name="controls_updated"),