+++ CHANGELOG.md
@@ -4,12 +4,24 @@ GovReady-Q Release Notes
 v999 (February XX, 2021)
 ------------------------
 
-[Add development changes here]
+Add System Assessment Report tracking to associate assessments and evidence with the system.
+Add initial dynamic status information to the project page.
+Project page displays mini-dashboard of compliance stats.
+
+**Feature changes**
+
+* Add System Assessment Report tracking to associate assessments and evidence with the system.
+* Add initial dynamic status information to the project page.
 
 **UI changes**
 
 * Improve page load times for listings with pagination and ordering for project listing and selected component listing. 
 * Display projects in pages of 10 and selected components by 5.
+* Project page displays mini-dashboard of compliance stats.
+    * Number of controls implemented out of count of controls.
+    * Number of POA&Ms.
+    * Count of system components.
+    * Approximate overall compliance based on controls implemented / count of controls.
 
 **Developer changes**
 
@@ -31,6 +43,8 @@ v0.9.1.51 (February 03, 2021)
 * Display components alphabetically in component library text listing and in selected components text listing.
 * Include a component description and statement count in component library text listing and in selected components text listing.
 * Remove admin's "update certified text" option from editing control implmentation statements.
+* New pages for System Assessment Report
+* Updates to project page for status information and other project information.
 
 **Developer changes**
 
@@ -82,6 +96,7 @@ Keeps track of app inputs by relating them to the app version.
 
 Includes the following schema update to the app.yaml file of Compliance Apps.
 Inputs are supported in the app.yaml file with the following format:
+
 ```
 input:
 - id: <input_id> (string)